modifed errors

+++ root-project/workers/unity/Assets/Scripts/AnimationRigging/AimSpeedController.cs
@@ -80,7 +80,7 @@ namespace AdvancedGears
             this.speed = speed;
         }
 
-        public void Rotate(double time)
+        public void Rotate(float time)
         {
             if (target == null)
                 return;
